refactor: Extract Welcome component in GetStarted page
