+++ packages/kit/src/views/Onboarding/pages/GetStarted/components/Welcome.tsx
@@ -1,0 +1,62 @@
+import { useIntl } from 'react-intl';
+
+import {
+  Heading,
+  Image,
+  LinearGradient,
+  SizableText,
+  Stack,
+  ThemeableStack,
+} from '@onekeyhq/components';
+import { MultipleClickStack } from '@onekeyhq/kit/src/components/MultipleClickStack';
+import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
+import { ETranslations } from '@onekeyhq/shared/src/locale';
+
+export function Welcome() {
+  const intl = useIntl();
+  const navigation = useAppNavigation();
+
+  return (
+    <Stack flex={1}>
+      <ThemeableStack fullscreen alignItems="center" justifyContent="center">
+        <MultipleClickStack
+          onPress={() => {
+            void navigation.popStack();
+          }}
+        >
+          <Image
+            w={360}
+            h={360}
+            source={require('@onekeyhq/kit/assets/logo-press.png')}
+          />
+        </MultipleClickStack>
+      </ThemeableStack>
+
+      <Stack px="$5" pt="$10" mt="auto">
+        <LinearGradient
+          position="absolute"
+          top="$0"
+          left="$0"
+          right="$0"
+          bottom="$0"
+          colors={['transparent', '$bgApp']}
+          $platform-native={{
+            display: 'none',
+          }}
+        />
+        <Stack zIndex={1}>
+          <Heading size="$heading4xl" textAlign="center">
+            {intl.formatMessage({
+              id: ETranslations.onboarding_welcome_message,
+            })}
+          </Heading>
+          <SizableText size="$bodyLg" textAlign="center" color="$textSubdued">
+            {intl.formatMessage({
+              id: ETranslations.onboarding_welcome_description,
+            })}
+          </SizableText>
+        </Stack>
+      </Stack>
+    </Stack>
+  );
+}
